revised train loading technique
In this alteration to the training script, instead of using nested for loops which both extract examples from the dataloader, which is somewhat confusing, I used a technique that converts the data loader to a python iterator that is shuffled at the start of each epoch. Two images are pulled from this shuffled iterator for the iteration through epoch steps. The potential advantage to this technique is that, as opposed to the nested loop strategy, both images in the pairing are new to the epoch. What I mean by that is: within the nested loop strat, at the start of each epoch, a single batch is pulled out that will be used in all the pairings against the batch pulled out at each step. I was thinking that this potentially would affect the distribution of the targets compared with the specific images they represent. This potential source of error is avoided using this iterator strategy, due to the shuffling.