improved garbage collection on data processing
There was an issue with running data processing in the same jupyter kernel as training, leading to rapid consumption of memory. Because of the way jupyter works, there is not nearly as much deallocation as in a python source file. I have moved the data prep process into a class within a seperate .py file so that all intermediate variables during the data prep process will be deallocated after calling the main() method.